Unify code between flattenInterfaceStructs and flattenHeaders
