+++ midend/flattenHeaders.h
@@ -20,29 +20,9 @@ limitations under the License.
 
 #include "ir/ir.h"
 #include "frontends/p4/typeChecking/typeChecker.h"
+#include "flattenInterfaceStructs.h"
 
 namespace P4 {
-
-/**
- * Policy to select which annotations of the nested struct to attach
- * to the struct fields after the nest struct is flattened.
- */
-class AnnotationSelectionPolicy {
- public:
-    virtual ~AnnotationSelectionPolicy() {}
-
-    /**
-     * Call for each nested struct to check if the annotation of the nested
-     * struct should be kept on its fields.
-     *
-     * @return
-     *  true if the annotation should be kept on the field.
-     *  false if the annotation should be discarded.
-     */
-    virtual bool keep(const IR::Annotation*) {
-        return false;
-    }
-};
 
 /**
 Find the types to replace and insert them in the nested struct map.
@@ -50,40 +30,7 @@ Find the types to replace and insert them in the nested struct map.
 class FindHeaderTypesToReplace : public Inspector {
     P4::TypeMap* typeMap;
     AnnotationSelectionPolicy *policy;
-
-    struct HeaderTypeReplacement : public IHasDbPrint {
-        HeaderTypeReplacement(const P4::TypeMap* typeMap, const IR::Type_Header* type,
-                AnnotationSelectionPolicy *policy);
-        const P4::TypeMap *typeMap;
-
-        // Maps nested field names to final field names.
-        // In our example this could be:
-        // .t.s.a -> _t_s_a0;
-        // .t.s.b -> _t_s_b1;
-        // .t.y -> _t_y2;
-        // .x -> _x3;
-        std::map<cstring, cstring> fieldNameRemap;
-        // Holds a new flat type
-        // struct M {
-        //    bit _t_s_a0;
-        //    bool _t_s_b1;
-        //    bit<6> _t_y2;
-        //    bit<3> _x3;
-        // }
-        const IR::Type* replacementType;
-        virtual void dbprint(std::ostream& out) const {
-            out << replacementType;
-        }
-
-        // Helper for constructor
-        void flatten(cstring prefix,
-                     const IR::Type* type,
-                     const IR::Annotations* annos,
-                     IR::IndexedVector<IR::StructField> *fields,
-                     AnnotationSelectionPolicy *policy);
-    };
-
-    ordered_map<cstring, HeaderTypeReplacement*> replacement;
+    ordered_map<cstring, StructTypeReplacement<IR::Type_StructLike>*> replacement;
 
  public:
     explicit FindHeaderTypesToReplace(P4::TypeMap *typeMap,
@@ -94,13 +41,14 @@ class FindHeaderTypesToReplace : public Inspector {
     }
     bool preorder(const IR::Type_Header* type) override;
     void createReplacement(const IR::Type_Header* type, AnnotationSelectionPolicy *policy);
-    HeaderTypeReplacement* getReplacement(const cstring name) const {
+    StructTypeReplacement<IR::Type_StructLike>* getReplacement(const cstring name) const {
         return ::get(replacement, name); }
     bool empty() const { return replacement.empty(); }
 };
 
 /**
- * This pass flattens any nested struct inside a P4 header
+ * This pass flattens any nested struct inside a P4 header.
+   This pass will fail if a header with nested fields is used as a left-value.
 
  * Should be run before the flattenInterfaceStructs pass.
 
@@ -132,7 +80,7 @@ struct local_metadata_t {
   bitvec_hdr bvh1;
 };
 
-The flattened data structures are shown below.  
+The flattened data structures are shown below.
 
 struct alt_t {
     bit<1> valid;
@@ -168,9 +116,9 @@ struct local_metadata_t {
     bitvec_hdr _bvh09;
     bitvec_hdr _bvh110;
 }
-	
+
 */
-class ReplaceHeaders : public Transform {
+class ReplaceHeaders : public Transform, P4WriteContext {
     P4::TypeMap *typeMap;
     FindHeaderTypesToReplace* findHeaderTypesToReplace;
